Remove hooks interfaces package
After multiple rewrites trying to find a set of interfaces to standardize
hooks, listeners and listeners removers I end up in understanding that
isn't possible since every system can implement his own proper concept
of dispatcher and listener.

That means system such as WordPress need to deal with the number of arguments
but other not, also may be there are system where there's no need for
listener priority etc...

Since isn't possible to standardize nor hook nor remove concept I think
it's better that the system define what a dispatcher get in and what is
needed in order to remove listeners.

So here we now have only invokable classes which get injected the callback
they wrap, additional configurations when needed and a callable which
will be called when we want to remove the listener, we just pass the
listener, then it is up to the system connect the listener to the event
and pass specific configurations.

Factories are no longer needed, this kind of invokable classes are
simply treated as newable objects.

If you need factories you may simply create a factory function within
a container, that's it, no need to have additional classes.

This make also much more simple to introduce a new listener.
